LayoutTests/platform/mac/accessibility/seamless-iframe.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script>
    var group1;
    function runTest() {
        if (window.accessibilityController) {
             group1 = accessibilityController.accessibleElementById("group1");
             shouldBe("group1.childAtIndex(0).role", "'AXRole: AXScrollArea'");
             shouldBe("group1.childAtIndex(0).childAtIndex(0).role", "'AXRole: AXGroup'");
             shouldBe("group1.childAtIndex(0).childAtIndex(0).childAtIndex(0).role", "'AXRole: AXButton'");
        }
        testRunner.notifyDone();
    }
</script>
</head>
<body id="body">

<div id="group1" tabindex="0">
<iframe onload="setTimeout('runTest()', 20);" seamless width=100 height=100 srcdoc="<div role='button'>button</div><p>test<p>test2<p>test3<p>test4"></iframe>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that seamless iframes do not appear as web areas to accessibility clients");

    testRunner.waitUntilDone();

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
