LayoutTests/platform/mac/accessibility/seamless-iframe-expected.txt


This tests that seamless iframes do not appear as web areas to accessibility clients

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS group1.childAtIndex(0).role is 'AXRole: AXScrollArea'
PASS group1.childAtIndex(0).childAtIndex(0).role is 'AXRole: AXGroup'
PASS group1.childAtIndex(0).childAtIndex(0).childAtIndex(0).role is 'AXRole: AXButton'
